semantically split functions in to their own files and organize under a directory.
